feat: remove deprecated models and tests; restructure Widget and Section into card directory
